# src/main/java/com/kshrd/assessment/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Assessment Management API",
    "description": "API for managing assessments, exams, sections, questions, and student assessments",
    "version": "v4"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v4/student-assessments/{assessmentId}/start": {
      "post": {
        "tags": [
          "Student Assessment Management"
        ],
        "summary": "Start assessment",
        "description": "Starts an assessment when a student begins taking it",
        "operationId": "startAssessment",
        "parameters": [
          {
            "name": "assessmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentAssessmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/student-assessments/submit": {
      "post": {
        "tags": [
          "Student Assessment Management"
        ],
        "summary": "Submit assessment",
        "description": "Submits a completed assessment with score and duration",
        "operationId": "submitAssessment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentAssessmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/student-assessments/submit-with-answers": {
      "post": {
        "tags": [
          "Student Assessment Management"
        ],
        "summary": "Submit assessment with answers",
        "description": "Submits a completed assessment along with all question answers",
        "operationId": "submitAssessmentWithAnswers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentAssessmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/student-assessments/grade": {
      "post": {
        "tags": [
          "Student Assessment Management"
        ],
        "summary": "Grade assessment",
        "description": "Grades a submitted assessment. Only accessible by teachers and admins",
        "operationId": "gradeAssessment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentAssessmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/student-assessments/assign": {
      "post": {
        "tags": [
          "Student Assessment Management"
        ],
        "summary": "Assign assessment to student",
        "description": "Assigns an assessment to a student. Only accessible by teachers and admins",
        "operationId": "assignAssessment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentAssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentAssessmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v4/student-assessments/{assessmentId}": {
      "get": {
        "tags": [
          "Student Assessment Management"
        ],
        "summary": "Get my assessment",
        "description": "Retrieves a specific assessment for the currently authenticated student",
        "operationId": "getMyAssessment",
        "parameters": [
          {
            "name": "assessmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentAssessmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/student-assessments/my-assessments": {
      "get": {
        "tags": [
          "Student Assessment Management"
        ],
        "summary": "Get my assessments",
        "description": "Retrieves all assessments assigned to the currently authenticated student with pagination support",
        "operationId": "getMyAssessments",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseStudentAssessmentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v4/answers/assessment/{assessmentId}": {
      "get": {
        "tags": [
          "Answer Management"
        ],
        "summary": "Get my answers for assessment",
        "description": "Retrieves all answers for the current student in a specific assessment with pagination support",
        "operationId": "getMyAnswersByAssessment",
        "parameters": [
          {
            "name": "assessmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseAnswerResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerRequest": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "answer": {
            "type": "object",
            "additionalProperties": {

            }
          },
          "score": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "assessmentId",
          "questionId"
        ]
      },
      "AnswerResponse": {
        "type": "object",
        "properties": {
          "answerId": {
            "type": "string",
            "format": "uuid"
          },
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "answer": {
            "type": "object",
            "additionalProperties": {

            }
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseStudentAssessmentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/StudentAssessmentResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "StudentAssessmentResponse": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "SUBMITTED",
              "GRADED",
              "EXPIRED"
            ]
          },
          "joinAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationInMinute": {
            "type": "number",
            "format": "double"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "gradingStatus": {
            "type": "string"
          },
          "totalScore": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "SubmitAssessmentRequest": {
        "type": "object",
        "properties": {
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "durationInMinute": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "assessmentId",
          "durationInMinute",
          "score"
        ]
      },
      "SubmitAnswersRequest": {
        "type": "object",
        "properties": {
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerRequest"
            }
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "durationInMinute": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "assessmentId",
          "durationInMinute",
          "score"
        ]
      },
      "GradeAssessmentRequest": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "gradingStatus": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "assessmentId",
          "gradingStatus",
          "score",
          "studentId"
        ]
      },
      "StudentAssessmentRequest": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "assessmentId",
          "studentId"
        ]
      },
...
      "ApiResponsePageResponseStudentAssessmentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseStudentAssessmentResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "PageResponseStudentAssessmentResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentAssessmentResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponsePageResponseAnswerResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseAnswerResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "PageResponseAnswerResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}